Rename Background View to VLinearGradient. Create ViewModifier
